Código reorganizado. Issue #17
