+++ cadastro/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+from .models import Campus
+from .models import Curso
+from .models import Papel
+
+admin.site.register(Campus)
+admin.site.register(Curso)
+admin.site.register(Papel)
