+++ cadastro/urls.py
@@ -5,9 +5,11 @@ from rest_framework_swagger.views import get_swagger_view
 from . import views
 
 router = routers.DefaultRouter()
+router.register(r'campus', views.CampusViewSet)
 router.register(r'curso', views.CursoViewSet)
+router.register(r'papeis', views.PapeisViewSet)
 
-schema_view = get_swagger_view(title='Curso API')
+schema_view = get_swagger_view(title='API V1')
 
 urlpatterns = [
     url(r'^$', schema_view),
