--- campus/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import Campus
-
-admin.site.register(Campus)
